frontend/requirements.txt
PyQt==5.15.11